Factor out runTransaction to common code
